requirements.txt
asgiref==3.8.1
Django==5.1.3
django-debug-toolbar==4.4.6
pillow==11.0.0
sqlparse==0.5.2
typing_extensions==4.12.2
